+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 ***
 
-All notable changes to `navindex/html-formatter` will be documented in this file.
+All notable changes to `balinomad/html-formatter` will be documented in this file.
+
+## 0.1.2 - 2025-07-03
+
+-   Namespace change from `Navindex\HtmlFormatter` to `BaliNomad\HtmlFormatter`.
+-   `HTMLFormatter` properly renamed to `HtmlFormatter` in README.md.
+
+## 0.1.1 - 2021-10-14
+
+-   Static code check fix.
 
 ## 0.1.0 - 2021-09-28